Start a Unity3D importer branch. This will eventually allow importing a BSP into Unity3D as a set of meshes through an editor script. It's already completely integrated with UnityEngine's Vector3 and Plane classes (this was incredibly easy to do, because there was so much feature parity between my classes and Unity's).
So far it will load a BSP into the proper data structures, and they are ready to be parsed into meshes and GameObjects. I only have a test case for 007 Nightfire right now, and it creates a prefab for each brush, completely loses its material reference, and takes about 10 minutes for a medium sized map. Improvements and fixes to follow.