Add support for Bezier patches in a few id Tech 3 engine games. Needs more research for some formats.
Fix a rare divide-by-zero problem when decompiling very small maps.

Fix decompiling maps marked as read-only in the filesystem.

Fix "textures/" in front of texture names in Tech 3 decompiles.

Fix an issue where a thread would not move on to the next job if the current one throws an exception.

A couple other things here and there, mostly coding style and syntax.