Sync work done from another project.
Major refactoring of Entity and Entities and be more C#-ized, and using LINQ and all that good stuff.

Add a callback delegate with references to the Entity and the GameObject generated for it. This allows game code to get this information from this code regardless of assembly, as long as this code is compiled first or concurrently.

Rework a couple things in how BSP Importer and BSP Loader work.